move test case main class to make it more easier to find
